chore(settings): move some files to store/selectors/settings [YTFRONT-4463]
