+++ packages/ui/src/ui/containers/SettingsPanel/settings-description.tsx
@@ -33,7 +33,7 @@ import ypath from '../../common/thor/ypath';
 import {AGGREGATOR_RADIO_ITEMS} from '../../constants/operations/statistics';
 import {NAMESPACES, SettingName} from '../../../shared/constants/settings';
 import {getRecentPagesInfo} from '../../store/selectors/slideoutMenu';
-import {getCurrentClusterNS} from '../../store/selectors/settings-ts';
+import {getCurrentClusterNS} from '../../store/selectors/settings/settings-ts';
 import SettingsMenuItem from '../../containers/SettingsMenu/SettingsMenuItem';
 import SettingsMenuRadio from '../../containers/SettingsMenu/SettingsMenuRadio';
 import SettingsMenuInput from '../SettingsMenu/SettingsMenuInput';
