+++ packages/ui/src/ui/pages/navigation/tabs/Queue/views/Partitions/ColumnsButton.connected.ts
@@ -4,7 +4,7 @@ import {connect} from 'react-redux';
 import ColumnsButton from '../../../../../../pages/navigation/tabs/Queue/ColumnsButton/ColumnsButton';
 import {setSettingsNavigationQueuePartitionsVisibility} from '../../../../../../store/actions/settings/settings';
 import type {RootState} from '../../../../../../store/reducers';
-import {getSettingsNavigationQueuePartitionsVisibility} from '../../../../../../store/selectors/settings-ts';
+import {getSettingsNavigationQueuePartitionsVisibility} from '../../../../../../store/selectors/settings/settings-ts';
 
 const allColumns = [
     {
